requirements.txt
python-dotenv>=1.0.0
openai>=1.2.0
PyAudio>=0.2.13
pydantic>=2.0.0
aiohttp>=3.8.5
asyncio>=3.4.3
wave>=0.0.2